+++ fe2-android/app/src/main/java/com/github/dedis/popstellar/ui/home/LaoCreateFragment.java
@@ -12,7 +12,6 @@ import androidx.fragment.app.Fragment;
 
 import com.github.dedis.popstellar.R;
 import com.github.dedis.popstellar.databinding.LaoCreateFragmentBinding;
-import com.github.dedis.popstellar.model.objects.security.Base64URLData;
 import com.github.dedis.popstellar.model.objects.security.PublicKey;
 import com.github.dedis.popstellar.repository.remote.GlobalNetworkManager;
 import com.github.dedis.popstellar.ui.lao.witness.WitnessingViewModel;
@@ -122,7 +121,7 @@ public final class LaoCreateFragment extends Fragment {
     // No need to have a LiveData as the fragment is recreated upon exiting the scanner
     List<String> witnesses =
         witnessingViewModel.getScannedWitnesses().stream()
-            .map(Base64URLData::getEncoded)
+            .map(PublicKey::getEncoded)
             .collect(Collectors.toList());
 
     // Show the witnesses title only if there's at least one witness
